moving ThreadCommands in dedicated module to avoid circular imports
