+++ src/pymodaq/utils/leco/daq_xDviewer_LECODirector.py
@@ -5,7 +5,7 @@ from typing import Optional, Union
 from easydict import EasyDict as edict
 
 from pymodaq.control_modules.viewer_utility_classes import DAQ_Viewer_base, comon_parameters, main
-from pymodaq.control_modules.utils import ThreadStatus, ThreadStatusViewer
+from pymodaq.control_modules.thread_commands import ThreadStatus, ThreadStatusViewer
 from pymodaq_utils.serialize.factory import SerializableFactory
 from pymodaq_utils.utils import ThreadCommand, getLineInfo
 
@@ -31,10 +31,6 @@ class DAQ_xDViewer_LECODirector(LECODirector, DAQ_Viewer_base):
 
     params_GRABBER = []
 
-    message_list = LECODirector.message_list + ["Quit", "Send Data 0D", "Send Data 1D",
-                                                "Send Data 2D", "Send Data ND",
-                                                "Status", "Done", "Server Closed",
-                                                "Info", "Infos", "Info_xml", 'x_axis', 'y_axis']
     socket_types = ["GRABBER"]
     params = comon_parameters + leco_parameters
     live_mode_available = True
